added dataset project config in BQ argument setter config
