+++ docs/BigQueryArgumentSetter-action.md
@@ -30,6 +30,10 @@
  that the BigQuery job will run in. If a temporary bucket needs to be created, the service account
  must have permission in this project to create buckets.
 
+**Dataset Project ID**: Project the dataset belongs to. This is only required if the dataset is not in 
+the same project that the BigQuery job will run in. If no value is given, it will default to the
+configured Project ID. 
+
  **Dataset**: Dataset the table belongs to. A dataset is contained within a specific project.
  Datasets are top-level containers that are used to organize and control access to tables and views.
