+++ src/main/java/io/cdap/plugin/gcp/bigquery/action/BigQueryExecute.java
@@ -181,15 +181,7 @@ public final class BigQueryExecute extends AbstractBigQueryAction {
     private static final String DATASET = "dataset";
     private static final String TABLE = "table";
     private static final String NAME_LOCATION = "location";
-    public static final String DATASET_PROJECT_ID = "datasetProject";
     private static final int ERROR_CODE_NOT_FOUND = 404;
-
-    @Name(DATASET_PROJECT_ID)
-    @Macro
-    @Nullable
-    @Description("The project in which the dataset specified in the `Dataset Name` is located or should be created."
-      + " Defaults to the project specified in the Project Id property.")
-    private String datasetProject;
 
     @Description("Dialect of the SQL command. The value must be 'legacy' or 'standard'. " +
       "If set to 'standard', the query will use BigQuery's standard SQL: " +
@@ -288,14 +280,6 @@ public final class BigQueryExecute extends AbstractBigQueryAction {
 
     public QueryJobConfiguration.Priority getMode() {
       return QueryJobConfiguration.Priority.valueOf(mode.toUpperCase());
-    }
-
-    @Nullable
-    public String getDatasetProject() {
-      if (GCPConfig.AUTO_DETECT.equalsIgnoreCase(datasetProject)) {
-        return ServiceOptions.getDefaultProjectId();
-      }
-      return Strings.isNullOrEmpty(datasetProject) ? getProject() : datasetProject;
     }
 
     @Nullable
