+++ widgets/BigQueryArgumentSetter-action.json
@@ -33,7 +33,15 @@
         },
         {
           "widget-type": "textbox",
-          "label": "Dataset",
+          "label": "Dataset Project ID",
+          "name": "datasetProject",
+          "widget-attributes": {
+            "placeholder": "The project in which the dataset specified in the `Dataset Name` is located or should be created. Defaults to the project specified in the Project Id property."
+          }
+        },
+        {
+          "widget-type": "textbox",
+          "label": "Dataset Name",
           "name": "dataset",
           "widget-attributes" : {
             "placeholder": "Dataset the table belongs to"
